fix multi-agent env top_down rendering bug in top_down_renderer.py
I found that in multi-agent envs, if render mode is top_down, the `position` attribute in `TopDownRenderer` class will be initialized as None. It's ok for the `_draw` method while in single-agent env, b/c the `self.current_track_vehicle` will have a value and is not None. But in MA Envs, since there is no `self.current_track_vehicle`, it will cause a bug of None Type error.  I kind of fix it following code in previous version of MetaDrive code. Hope it will help!