fix a rendering bug for top_down mode in multi-agent envs
I found that in multi-agent envs, if render mode is top_down, it will fail rendering on my Mac. I figure out that the `position` attribute in `TopDownRenderer` class will be initialized as None. It's ok for the `_draw` method while in single-agent env, b/c the `self.current_track_vehicle` will have a value and is not None. But in MA Envs, since there is no `self.current_track_vehicle`, it will cause a bug of None Type error.  I kind of fix it by following code in previous version of MetaDrive repo. Hope it will help!